Move traits to new module and clean module namespaces
