+++ src/airspace/conversion.rs
@@ -2,7 +2,7 @@ use std::convert::Infallible;
 use std::time::SystemTime;
 
 use crate::airspace::detail::AirspaceUpdate;
-use crate::core::central_disk_logger::interface::IntoLogMessage;
+use crate::core::central_disk_logger::IntoLogMessage;
 use crate::pb::airspace::PbAirspaceUpdate;
 
 impl IntoLogMessage<PbAirspaceUpdate> for AirspaceUpdate {
