+++ src/core/central_disk_logger/mod.rs
@@ -1,12 +1,16 @@
 pub mod errors;
 pub mod interface;
 pub mod task;
+pub mod traits;
 
 #[cfg(test)]
 pub mod testing;
 
+pub use errors::{
+    CentralDiskLoggerError, DiskloggerRegistryError, JsonLoggingError, ProtoLoggingError,
+};
 pub use interface::{
-    ChannelID, DiskLoggerMessage, DiskLoggerRegistry, JsonlLoggerHandle, LogSender,
-    ProtoLoggerHandle,
+    ChannelID, DiskLoggerMessage, DiskLoggerRegistry, JsonlLoggerHandle, ProtoLoggerHandle,
 };
 pub use task::CentralDiskLogger;
+pub use traits::{IntoLogMessage, LogSender};
